Presets page shows selected preset workout content
A new, react Preset interface was created to hold its necessary string and number data with useEffect imported from React to aid the side effect processing in PresetsList.tsx by asynchronously passing the default workout plan examples: Beginner, Intermediate, and Advanced to the presets state variable. Its optional dependency was the empty array ‘[]’ for now since it doesn’t need to re-render for any changes because the backend/api isn’t fully connected to more workout plans being queried from the database. Side-by-side, the presets-list and selectedpreset-contents are returned. The logic for the presets-list column is similar with the checked and the onChange reflecting the state of the selected preset as it maps and iterates through plan_name variables. If there’s a selected preset, the days_of_the_week strings are printed out continually with a comma and space delimiting, and then it maps through the index of the array holding workouts to print out the name, sets, reps, and weights of each in unordered list format.